moving modals to simplify app component
